requirements.txt
mediapipe==0.10.21
numpy>=1.26.0
opencv-python
pyttsx3
Pillow
